light ui changes, added conditional rendering of submitting food items.
