+++ client/src/components/dashboard/Daily.tsx
@@ -1,7 +1,6 @@
 
 import React, { useEffect } from "react";
-import Box from "@mui/material/Box";
-import Typography from "@mui/material/Typography";
+import { Box, Typography, Button } from "@mui/material";
 import CalorieBar from "./calorieBar";
 import ExceedBar from './exceedBar';
 import './daily.scss';
@@ -11,7 +10,7 @@ const defaultWaterGoal = 8;
 
 const Daily = (props: any) => {
 
-  const { handleDailyUpdate, dailyData, userInfo } = props;
+  const { handleDailyUpdate, dailyData, userInfo, submitModalOn, toggleSubmit } = props;
   const calorieGoal = userInfo?.targetCalories || defaultCalorieGoal;
   const calorie = dailyData?.caloriesAmount; // will be changed;
   const waterGoal = userInfo?.targetWater || defaultWaterGoal;
@@ -28,7 +27,7 @@ const Daily = (props: any) => {
         <Box className="chart_item">
           <Typography className="title">Calories</Typography>
           <Box className="chart_wrap">
-          <Typography className="title">{`${calorie} / ${calorieGoal}`}</Typography>
+            <Typography className="title">{`${calorie} / ${calorieGoal}`}</Typography>
             <CalorieBar
               data={[{ value: calorie, goal: calorieGoal }]} />
           </Box>
@@ -36,10 +35,20 @@ const Daily = (props: any) => {
         <Box className="chart_item">
           <Typography className="title">Water</Typography>
           <Box className="chart_wrap">
-          <Typography className="title">{`${water} / ${waterGoal}  \t cups`}</Typography>
+            <Typography className="title">{`${water} / ${waterGoal}  \t cups`}</Typography>
             <ExceedBar
               data={[{ value: water, goal: waterGoal }]} />
           </Box>
+        </Box>
+        <Box>
+          {
+            submitModalOn ? null :
+              <Button sx={{width: '100%'}}
+                variant="contained"
+                onClick={() => { toggleSubmit(true) }} >
+                + Add Food Item
+              </Button>
+          }
         </Box>
       </Box>
     )
